refactor: 재사용을 위해 ReferenceLinkBox로 분할 및 LinkIconBox 이름 LinkItem으로 수정
