+++ src/components/molecules/ReferenceLinkBox/index.ts
@@ -1,0 +1,3 @@
+export { default as ReferenceLinkBox } from './ReferenceLinkBox';
+export { renderIcon } from './ReferenceLinkBox';
+export { LINK_ICON_TYPES } from './ReferenceLinkBox';
