+++ src/components/organisms/ResumeBasicInput/ReferenceLinkForm.tsx
@@ -1,7 +1,7 @@
 import { Box, Flex, FormControl, Radio, RadioGroup, Stack } from '@chakra-ui/react';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
-import LinkIconBox from './LinkIconBox';
+import LinkItem from './LinkItem';
 import { BorderBox } from '~/components/atoms/BorderBox';
 import { Button } from '~/components/atoms/Button';
 import { FormTextInput } from '~/components/molecules/FormTextInput';
@@ -15,7 +15,6 @@ const ReferenceLinkForm = () => {
   const [linkData, setLinkData] = useState<LinkData[]>([]);
 
   const handleRemoveLink = (urlToRemove: string) => {
-    // urlToRemove와 일치하지 않는 링크만 남기고 나머지를 필터링합니다.
     const updatedLinks = linkData.filter((link) => link.url !== urlToRemove);
     setLinkData(updatedLinks);
   };
@@ -48,7 +47,7 @@ const ReferenceLinkForm = () => {
           key={index}
           mb={3}
         >
-          <LinkIconBox
+          <LinkItem
             key={index}
             url={data.url}
             type={data.type}
